requirements.txt
fastapi==0.75.0
pydantic==1.9.0
starlette==0.17.1
uvicorn==0.17.5
asyncpg==0.25.0
sqlalchemy==1.4.31
loguru==0.6.0
alembic==1.7.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
httpx==0.22.0
xmltodict==0.12.0
aioredis==1.3.1
arq==0.22
requests==2.27.1